feat: Apply authorization for 'User' model

+++ app/Http/Controllers/UserController.php
@@ -2,46 +2,35 @@
 
 namespace App\Http\Controllers;
 
+use App\Actions\Users\QueryUsersAction;
+use App\Actions\Users\UpdateUserAction;
+use App\Http\Requests\Tasks\UpdateTaskRequest;
+use App\Http\Requests\Users\QueryUsersRequest;
 use App\Models\User;
-use Illuminate\Http\Request;
+use Gate;
 use Illuminate\Support\Facades\Auth;
-use function Pest\Laravel\json;
 
 class UserController extends Controller
 {
-    public function index(Request $request)
+    public function index(QueryUsersRequest $request)
     {
-        $users = User::query()
-            ->when($request->username, function ($query, $username) {
-                return $query->whereUsername($username);
-            })->when($request->id, function ($query, $id) {
-                $query->where('id', '=', $id);
-            })->when($request->before, function ($query, $before) {
-                $query->where('created_at', '<=', $before);
-            })->when($request->after, function ($query, $after) {
-                $query->where('created_at', '>=', $after);
-            })
-        ->get();
+        $users = QueryUsersAction::do($request);
 
         return view('users.index', ['users' => $users]);
     }
 
     public function show(User $user)
     {
+        Gate::authorize('vudd', $user);
+
         return view('users.show', ['user' => $user]);
     }
 
-    public function update(User $user, Request $request)
+    public function update(User $user, UpdateTaskRequest $request)
     {
-        $user->email = $request->email ?? $user->email;
+        Gate::authorize('vudd', $user);
 
-        $user->username = $request->username ?? $user->username;
-
-        if (!is_null($request->new_password)) {
-            $user->password = $request->new_password;
-        }
-
-        $user->save();
+        UpdateUserAction::do($user, $request);
 
         if ($user->is(Auth::user())) {
             Auth::logout();
@@ -53,14 +42,19 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize('vudd', $user);
+
         $user->delete();
     }
 
     public function dashboard(User $user)
     {
-        return $user->is_admin ?
-            view('users.admin.dashboard')
-            :
-            view('users.non_admin.dashboard', ['tasks' => $user->tasks]);
+        if ($user->is_admin) {
+            return view('users.admin.dashboard');
+        } else {
+            Gate::authorize('vudd', $user);
+
+            return view('users.non_admin.dashboard', ['tasks' => $user->tasks]);
+        }
     }
 }
